springboot handson

+++ 839290_Manish/Morning/SpringBoot/SpringBootApplication.zip_expanded/SpringBootApplication/src/main/java/com/org/SpringBootApplication/Service/CustomerService.java
@@ -8,6 +8,8 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import com.org.SpringBootApplication.Entities.Customer;
+import com.org.SpringBootApplication.Exception.CustomerAlreadyExistsException;
+import com.org.SpringBootApplication.Exception.CustomerNotFoundException;
 import com.org.SpringBootApplication.Repositories.CustomerRepository;
 
 @Service
@@ -17,34 +19,56 @@ public class CustomerService {
 	@Autowired
 	private CustomerRepository dao;
 	
-	public Customer addCustomer(Customer customer) {
+	public Customer addCustomer(Customer customer) throws CustomerAlreadyExistsException{
+		
+		List<Customer> customers = dao.findAll();
+		for(Customer customerold: customers){
+			if(customer.getName().equals(customerold.getName()))
+				throw new CustomerAlreadyExistsException("User name already Exists... You can try with another username");
+		}
 		Customer createdCustomer = dao.save(customer);
 		return createdCustomer;
+	
 	}
 	
-	public List<Customer> getAllCustomers(){
-		return dao.findAll();
+	public List<Customer> getAllCustomers() throws CustomerNotFoundException{
+		if(dao.findAll().isEmpty())
+			throw new CustomerNotFoundException("Customer data is empty");
+		else
+			return dao.findAll();
 	}
 	
-	public Customer getCustomerById(int id)
+	public Customer getCustomerById(int id) throws CustomerNotFoundException
 	{
 		Optional<Customer> option = dao.findById(id);
-		return option.orElse(null);
+		if(option.isPresent())
+		{
+				return option.get();
+		}
+		else
+		{
+			throw new CustomerNotFoundException("Sorry Customer with id number "+id+" not found");
+		}
 	}
 	
-	public Customer updateCustomer(int id, LocalDate dob) {
+	public Customer updateCustomer(int id, LocalDate dob) throws CustomerNotFoundException {
 		Customer customer = getCustomerById(id);
 		if(customer != null) {
 			customer.setDob(dob);
 			return dao.save(customer);
 		} 
 		else 
-			return null;
+			throw new CustomerNotFoundException("Sorry Customer with id number "+id+" not found");
 	}
 	
-	public void deleteCustomerById(int id)
+	public void deleteCustomerById(int id) throws CustomerNotFoundException
 	{
-		dao.deleteById(id);
+		Customer customer = getCustomerById(id);
+		if(customer != null) {
+			dao.deleteById(id);
+		} 
+		else 
+			throw new CustomerNotFoundException("Sorry Customer with id number "+id+" not found");
 	}
 
 }
